requirements.txt
aiosqlite==0.20.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
asyncpg==0.30.0
attrs==24.3.0
babel==2.16.0
bcrypt==4.2.1
bibtexparser==2.0.0b8
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-spinner==0.1.10
clldutils==3.24.0
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.1
croniter==1.4.1
csvw==3.5.1
cycler==0.12.1
dlinfo==2.0.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.6
fastapi-cli==0.0.7
fastapi-utilities==0.3.0
filelock==3.13.1
fonttools==4.55.3
fpdf==1.7.2
fsspec==2024.2.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
iniconfig==2.0.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
language-tags==1.2.0
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mpmath==1.3.0
munch==4.0.0
narwhals==1.22.0
networkx==3.2.1
numpy==2.2.1
orjson==3.10.14
packaging==24.2
pandas==2.2.3
passlib==1.7.4
phonemizer==3.3.0
pillow==11.1.0
pluggy==1.5.0
protobuf==5.29.3
pyarrow==18.1.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.5
pydantic-extra-types==2.10.1
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
pylatexenc==2.10
pyparsing==3.2.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
python-jose==3.3.0
python-multipart==0.0.20
python-telegram-bot==21.10
pytz==2024.2
PyYAML==6.0.2
rdflib==7.1.3
referencing==0.35.1
regex==2024.11.6
reportlab==4.2.5
requests==2.32.3
rfc3986==1.5.0
rich==13.9.4
rich-toolkit==0.13.2
rpds-py==0.22.3
rsa==4.9
safetensors==0.5.2
scipy==1.15.1
seaborn==0.13.2
segments==2.2.1
sendgrid==6.11.0
setuptools==70.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.0
SQLAlchemy==2.0.37
starkbank-ecdsa==2.2.0
starlette==0.41.3
streamlit==1.41.1
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tokenizers==0.21.0
toml==0.10.2
torch==2.5.1+cpu
torchaudio==2.5.1+cpu
torchvision==0.20.1+cpu
tornado==6.4.2
tqdm==4.67.1
transformers==4.48.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.0.4
websockets==14.1
